chore: organize type files - Increases stringency.
